one more fix for the receipts
