+++ subgraph/src/public-lock.ts
@@ -1,11 +1,4 @@
-import {
-  Address,
-  BigInt,
-  log,
-  Bytes,
-  ethereum,
-  store,
-} from '@graphprotocol/graph-ts'
+import { Address, BigInt, log, Bytes, store } from '@graphprotocol/graph-ts'
 import {
   CancelKey as CancelKeyEvent,
   ExpirationChanged as ExpirationChangedUntilV11Event,
@@ -24,7 +17,7 @@ import {
 } from '../generated/templates/PublicLock/PublicLock'
 
 import { PublicLockV11 as PublicLock } from '../generated/templates/PublicLock/PublicLockV11'
-import { Key, Lock, UnlockStats, LockStats, Receipt } from '../generated/schema'
+import { Key, Lock, UnlockStats, LockStats } from '../generated/schema'
 
 import {
   genKeyID,
@@ -33,7 +26,7 @@ import {
   getKeyManagerOf,
   LOCK_MANAGER,
 } from './helpers'
-import { nullAddress } from '../tests/constants'
+import { createReceipt } from './receipt'
 
 function newKey(event: TransferEvent): void {
   const keyID = genKeyID(event.address, event.params.tokenId.toString())
@@ -75,7 +68,7 @@ function newKey(event: TransferEvent): void {
 
   key.save()
 
-  // create receipt
+  // creat
   createReceipt(event)
 
   // update lock
@@ -405,81 +398,5 @@ export function handleLockMetadata(event: LockMetadataEvent): void {
 
     // lock.symbol = event.params.symbol
     lock.save()
-  }
-}
-
-/**
- * Create Receipt object for subgraph for key 'purchase'/'extend'/'renewal'
- * @param {String} keyID - key id
- * @param {event} Object - Object event
- * @return {void}
- */
-export function createReceipt(event: ethereum.Event): void {
-  const lockAddress = event.address.toHexString()
-  const hash = event.transaction.hash.toHexString()
-
-  const receipt = new Receipt(hash)
-
-  const lock = Lock.load(lockAddress)
-
-  const tokenAddress =
-    lock && lock.tokenAddress
-      ? lock.tokenAddress
-      : Bytes.fromHexString(nullAddress)
-
-  // TODO: compile from contract ABI
-  // WARNING : For some tokens it may be different. In that case we would move to a list!
-  const ERC20_TRANSFER_TOPIC0 =
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef'
-
-  if (tokenAddress != Bytes.fromHexString(nullAddress)) {
-    const txReceipt = event.receipt!
-    const logs: ethereum.Log[] = txReceipt.logs
-    if (logs) {
-      // If it is an ERC20 lock, there should be multiple events
-      // including one for the ERC20 transfer
-      for (let i = 0; i < logs.length; i++) {
-        const txLog = logs[i]
-        if (
-          txLog.address == tokenAddress &&
-          // Do we always have txLog.topics[0] ?
-          txLog.topics[0].toHexString() == ERC20_TRANSFER_TOPIC0
-        ) {
-          receipt.payer = ethereum
-            .decode('address', txLog.topics[1])!
-            .toAddress()
-            .toHexString()
-
-          receipt.amountTransferred = ethereum
-            .decode('uint256', txLog.data)!
-            .toBigInt()
-        }
-      }
-      // If no ERC20 transfer event was found, this was not a "paid" transaction,
-      // which means we don't need to create a receipt.
-    }
-  } else {
-    receipt.payer = event.transaction.from.toHexString()
-    receipt.amountTransferred = event.transaction.value
-  }
-
-  const totalGas = event.transaction.gasPrice.plus(event.transaction.gasLimit)
-  receipt.lockAddress = lockAddress
-  receipt.timestamp = event.block.timestamp
-  receipt.sender = event.transaction.from.toHexString()
-  receipt.tokenAddress = tokenAddress.toHexString()
-  receipt.gasTotal = BigInt.fromString(totalGas.toString())
-
-  if (lock) {
-    const newReceiptNumber = lock.numberOfReceipts.plus(BigInt.fromI32(1))
-    lock.numberOfReceipts = newReceiptNumber
-    receipt.numberOfReceipt = newReceiptNumber
-    lock.save()
-  }
-
-  // save receipt, but only if we have a payer
-  // (i.e. this is a paid transaction)
-  if (receipt.payer && receipt.amountTransferred > BigInt.fromI32(0)) {
-    receipt.save()
   }
 }
