+++ subgraph/tests/keys-utils.ts
@@ -265,6 +265,7 @@ export function createKeyExtendedEvent(
 
   return keyExtendedEvent
 }
+
 // before v10
 export function createRenewKeyPurchaseEvent(
   owner: Address,
@@ -273,6 +274,7 @@ export function createRenewKeyPurchaseEvent(
   const renewKeyPurchaseEvent = changetype<RenewKeyPurchase>(newMockEvent())
 
   renewKeyPurchaseEvent.address = dataSource.address()
+  log.info(dataSource.address().toHexString(), [])
 
   renewKeyPurchaseEvent.parameters = []
